Fixing compiler warnings when using explicit integer types with generalised printf() templates in debug output.

+++ stackgvm/source/include/gvm_branch/bit_set.hpp
@@ -2,11 +2,11 @@
 // Literal bit set in Local
 BCC(X_BS_SL) {
     gvmDebugOpcode(
-        "bcc set #%d, (%d), %d : 0x%08X => ",
-        U8(1),
-        OPS(2),
+        "bcc set #%u, (%d), %d : 0x%08X => ",
+        (unsigned)U8(1),
+        (int)OPS(2),
         (int)J16(3),
-        LOC(2).u
+        (unsigned)LOC(2).u
     );
     if ((1<<U8(1)) & LOC(2).u) {
         gvmDebugJump(3);
@@ -21,11 +21,11 @@ BCC(X_BS_SL) {
 // Literal bit set in Indirect[0]
 BCC(X_BS_SI0) {
     gvmDebugOpcode(
-        "bcc set #%d, (i0 + %u), %d : 0x%08X => ",
-        U8(1),
-        OPU(2),
+        "bcc set #%u, (i0 + %u), %d : 0x%08X => ",
+        (unsigned)U8(1),
+        (unsigned)OPU(2),
         (int)J16(3),
-        IX0(2).u
+        (unsigned)IX0(2).u
     );
     if ((1<<U8(1)) & IX0(2).u) {
         gvmDebugJump(3);
@@ -40,11 +40,11 @@ BCC(X_BS_SI0) {
 // Literal bit set in Indirect[1]
 BCC(X_BS_SI1) {
     gvmDebugOpcode(
-        "bcc set #%d, (i1 + %u), %d : 0x%08X => ",
-        U8(1),
-        OPU(2),
+        "bcc set #%u, (i1 + %u), %d : 0x%08X => ",
+        (unsigned)U8(1),
+        (unsigned)OPU(2),
         (int)J16(3),
-        IX1(2).u
+        (unsigned)IX1(2).u
     );
     if ((1<<U8(1)) & IX1(2).u) {
         gvmDebugJump(3);
@@ -60,10 +60,10 @@ BCC(X_BS_SI1) {
 BCC(X_BS_LL) {
     gvmDebugOpcode(
         "bcc set (%d), (%d), %d : 0x%08X => ",
-        OPS(1),
-        OPS(2),
+        (int)OPS(1),
+        (int)OPS(2),
         (int)J16(3),
-        LOC(2).u
+        (unsigned)LOC(2).u
     );
     if ((1<<LOC(1).u) & LOC(2).u) {
         gvmDebugJump(3);
@@ -79,10 +79,10 @@ BCC(X_BS_LL) {
 BCC(X_BS_LI0) {
     gvmDebugOpcode(
         "bcc set (%d), (i0 + %u), %d : 0x%08X => ",
-        OPS(1),
-        OPU(2),
+        (int)OPS(1),
+        (unsigned)OPU(2),
         (int)J16(3),
-        IX0(2).u
+        (unsigned)IX0(2).u
     );
     if ((1<<LOC(1).u) & IX0(2).u) {
         gvmDebugJump(3);
@@ -98,10 +98,10 @@ BCC(X_BS_LI0) {
 BCC(X_BS_LI1) {
     gvmDebugOpcode(
         "bcc set (%d), (i1 + %u), %d : 0x%08X => ",
-        OPS(1),
-        OPU(2),
+        (int)OPS(1),
+        (unsigned)OPU(2),
         (int)J16(3),
-        IX1(2).u
+        (unsigned)IX1(2).u
     );
     if ((1<<LOC(1).u) & IX1(2).u) {
         gvmDebugJump(3);
