+++ stackgvm/source/include/gvm_branch/int_less_or_equal.hpp
@@ -3,11 +3,11 @@
 BCC(I_LE_LS) {
     gvmDebugOpcode(
         "bcc le.i (%d), #%d, %d : %d <= %d => ",
-        OPS(1),
-        S8(2),
+        (int)OPS(1),
+        (int)S8(2),
         (int)J16(3),
-        LOC(1).i,
-        S8(2)
+        (int)LOC(1).i,
+        (int)S8(2)
     );
     if (LOC(1).i <= S8(2)) {
         gvmDebugJump(3);
@@ -23,11 +23,11 @@ BCC(I_LE_LS) {
 BCC(I_LE_I0S) {
     gvmDebugOpcode(
         "bcc le.i (i0 + %u), #%d, %d : %d <= %d => ",
-        OPU(1),
-        S8(2),
+        (unsigned)OPU(1),
+        (int)S8(2),
         (int)J16(3),
-        IX0(1).i,
-        S8(2)
+        (int)IX0(1).i,
+        (int)S8(2)
     );
     if (IX0(1).i <= S8(2)) {
         gvmDebugJump(3);
@@ -43,11 +43,11 @@ BCC(I_LE_I0S) {
 BCC(I_LE_I1S) {
     gvmDebugOpcode(
         "bcc le.i (i1 + %u), #%d, %d : %d <= %d => ",
-        OPU(1),
-        S8(2),
+        (unsigned)OPU(1),
+        (int)S8(2),
         (int)J16(3),
-        IX1(1).i,
-        S8(2)
+        (int)IX1(1).i,
+        (int)S8(2)
     );
     if (IX1(1).i <= S8(2)) {
         gvmDebugJump(3);
@@ -63,11 +63,11 @@ BCC(I_LE_I1S) {
 BCC(I_LE_LL) {
     gvmDebugOpcode(
         "bcc le.i (%d), (%d), %d : %d <= %d => ",
-        OPS(1),
-        OPS(2),
+        (int)OPS(1),
+        (int)OPS(2),
         (int)J16(3),
-        LOC(1).i,
-        LOC(2).i
+        (int)LOC(1).i,
+        (int)LOC(2).i
     );
     if (LOC(1).i <= LOC(2).i) {
         gvmDebugJump(3);
@@ -84,10 +84,10 @@ BCC(I_LE_LI0) {
     gvmDebugOpcode(
         "bcc le.i (%d), (i0 + %u), %d : %d <= %d => ",
         OPS(1),
-        OPU(2),
+        (unsigned)OPU(2),
         (int)J16(3),
-        LOC(1).i,
-        IX0(2).i
+        (int)LOC(1).i,
+        (int)IX0(2).i
     );
     if (LOC(1).i <= IX0(2).i) {
         gvmDebugJump(3);
@@ -104,10 +104,10 @@ BCC(I_LE_LI1) {
     gvmDebugOpcode(
         "bcc le.i (%d), (i1 + %u), %d : %d <= %d => ",
         OPS(1),
-        OPU(2),
+        (unsigned)OPU(2),
         (int)J16(3),
-        LOC(1).i,
-        IX1(2).i
+        (int)LOC(1).i,
+        (int)IX1(2).i
     );
     if (LOC(1).i <= IX1(2).i) {
         gvmDebugJump(3);
@@ -123,11 +123,11 @@ BCC(I_LE_LI1) {
 BCC(I_LE_II) {
     gvmDebugOpcode(
         "bcc le.i (i0 + %u), (i1 + %u), %d : %d <= %d => ",
-        OPU(1),
-        OPU(2),
+        (unsigned)OPU(1),
+        (unsigned)OPU(2),
         (int)J16(3),
-        IX0(1).i,
-        IX1(2).i
+        (int)IX0(1).i,
+        (int)IX1(2).i
     );
     if (IX0(1).i <= IX1(2).i) {
         gvmDebugJump(3);
